+++ src/Lib/Simplify/Exception/Simplify_AuthenticationException.php
@@ -1,0 +1,15 @@
+<?
+
+namespace Lib\Simplify\Exception;
+
+use Lib\Simplify\Exception\Simplify_ApiException;
+
+class Simplify_AuthenticationException extends Simplify_ApiException {
+
+    /**
+     * @ignore
+     */
+    function __construct($message, $status = null, $errorData = null) {
+        parent::__construct($message, $status, $errorData);
+    }
+}
