+++ src/Lib/Simplify/Simplify_Object.php
@@ -1,6 +1,8 @@
 <?php
+
+namespace Lib\Simplify;
 /*
- * Copyright (c) 2013 - 2018 MasterCard International Incorporated
+ * Copyright (c) 2013 - 2019 MasterCard International Incorporated
  * All rights reserved.
  * 
  * Redistribution and use in source and binary forms, with or without modification, are
