+++ src/Lib/Simplify/Simplify_PaymentsApi.php
@@ -1,6 +1,14 @@
 <?php
+
+namespace Lib\Simplify;
+
+use Lib\Simplify;
+use Lib\Simplify\Simplify_HTTP;
+use Lib\Simplify\Simplify_PaymentsApi;
+use Lib\Simplify\Simplify_ResourceList;
+use Lib\Simplify\Simplify_Authentication;
 /*
- * Copyright (c) 2013 - 2018 MasterCard International Incorporated
+ * Copyright (c) 2013 - 2019 MasterCard International Incorporated
  * All rights reserved.
  * 
  * Redistribution and use in source and binary forms, with or without modification, are
