+++ src/Lib/Simplify/Simplify_ResourceList.php
@@ -1,6 +1,8 @@
 <?php
+
+namespace Lib\Simplify;
 /*
- * Copyright (c) 2013 - 2018 MasterCard International Incorporated
+ * Copyright (c) 2013 - 2019 MasterCard International Incorporated
  * All rights reserved.
  * 
  * Redistribution and use in source and binary forms, with or without modification, are 
@@ -27,38 +29,18 @@
  */
 
 
-class Simplify_Event extends Simplify_Object {
+/**
+ * Class Simplify_ResourceList represents a collection of domain objects returned by one of the list<Domain>() methods.
+ */
+class Simplify_ResourceList {
 
     /**
-     * Creates an Event object
-     * @param     array $hash A map of parameters; valid keys are:
-     *     <dt><code>paylod</code></dt>    <dd>The raw JWS payload. </dd> <strong>required</strong>
-     *     <dt><code>url</code></dt>    <dd>The URL for the webhook.  If present it must match the URL registered for the webhook.</dd>
-     * @param  $authentication Object that contains the API public and private keys.  If null the values of the static
-     *         Simplify::$publicKey and Simplify::$privateKey will be used.
-     * @return Payments_Event an Event object.
-     * @throws InvalidArgumentException
+     * @var array $list the list of domain objects.
      */
-    static public function createEvent($hash, $authentication = null) {
-
-        $args = func_get_args();
-        $authentication = Simplify_PaymentsApi::buildAuthenticationObject($authentication, $args, 2);
-
-        $paymentsApi = new Simplify_PaymentsApi();
-
-        $jsonObject = $paymentsApi->jwsDecode($hash, $authentication);
-
-        if ($jsonObject['event'] == null) {
-            throw new InvalidArgumentException("Incorect data in webhook event");
-        }   
-
-        return  $paymentsApi->convertFromHashToObject($jsonObject['event'], self::getClazz());
-    }
+    public $list = array();
 
     /**
-     * @ignore
+     * @var int $total the total number of object available.
      */
-    static public function getClazz() {
-        return "Event";
-    }
+    public $total = 0;
 }
