+++ src/Lib/Simplify/Simplify_Tax.php
@@ -1,6 +1,12 @@
 <?php
+
+namespace Lib\Simplify;
+
+use Lib\Simplify\Simplify_Tax;
+use Lib\Simplify\Simplify_Object;
+use Lib\Simplify\Simplify_PaymentsApi;
 /*
- * Copyright (c) 2013 - 2018 MasterCard International Incorporated
+ * Copyright (c) 2013 - 2019 MasterCard International Incorporated
  * All rights reserved.
  * 
  * Redistribution and use in source and binary forms, with or without modification, are 
@@ -27,20 +33,21 @@
  */
 
 
-class Simplify_Webhook extends Simplify_Object {
+class Simplify_Tax extends Simplify_Object {
     /**
-     * Creates an Simplify_Webhook object
+     * Creates an Simplify_Tax object
      * @param     array $hash a map of parameters; valid keys are:<dl style="padding-left:10px;">
-     *     <dt><tt>url</tt></dt>    <dd>Endpoint URL <strong>required </strong></dd></dl>
+     *     <dt><tt>label</tt></dt>    <dd>The label of the tax object. [max length: 255] <strong>required </strong></dd>
+     *     <dt><tt>rate</tt></dt>    <dd>The tax rate.  Decimal value up three decimal places.  e.g 12.501. [max length: 6] <strong>required </strong></dd></dl>
      * @param     $authentication -  information used for the API call.  If no value is passed the global keys Simplify::public_key and Simplify::private_key are used.  <i>For backwards compatibility the public and private keys may be passed instead of the authentication object.<i/>
-     * @return    Webhook a Webhook object.
+     * @return    Tax a Tax object.
      */
-    static public function createWebhook($hash, $authentication = null) {
+    static public function createTax($hash, $authentication = null) {
 
         $args = func_get_args();
         $authentication = Simplify_PaymentsApi::buildAuthenticationObject($authentication, $args, 2);
 
-        $instance = new Simplify_Webhook();
+        $instance = new Simplify_Tax();
         $instance->setAll($hash);
 
         $object = Simplify_PaymentsApi::createObject($instance, $authentication);
@@ -51,11 +58,11 @@ class Simplify_Webhook extends Simplify_Object {
 
 
        /**
-        * Deletes an Simplify_Webhook object.
+        * Deletes an Simplify_Tax object.
         *
         * @param     $authentication -  information used for the API call.  If no value is passed the global keys Simplify::public_key and Simplify::private_key are used.  <i>For backwards compatibility the public and private keys may be passed instead of the authentication object.</i>
         */
-        public function deleteWebhook($authentication = null) {
+        public function deleteTax($authentication = null) {
 
             $args = func_get_args();
             $authentication = Simplify_PaymentsApi::buildAuthenticationObject($authentication, $args, 1);
@@ -67,23 +74,23 @@ class Simplify_Webhook extends Simplify_Object {
 
 
        /**
-        * Retrieve Simplify_Webhook objects.
+        * Retrieve Simplify_Tax objects.
         * @param     array criteria a map of parameters; valid keys are:<dl style="padding-left:10px;">
-        *     <dt><tt>filter</tt></dt>    <dd>Filters to apply to the list.  </dd>
+        *     <dt><tt>filter</tt></dt>    <dd><table class="filter_list"><tr><td>filter.id</td><td>Filter by the Id of the tax</td></tr><tr><td>filter.label</td><td>Filter by the label(name) of the tax</td></tr></table>  </dd>
         *     <dt><tt>max</tt></dt>    <dd>Allows up to a max of 50 list items to return. [min value: 0, max value: 50, default: 20]  </dd>
         *     <dt><tt>offset</tt></dt>    <dd>Used in paging of the list.  This is the start offset of the page. [min value: 0, default: 0]  </dd>
-        *     <dt><tt>sorting</tt></dt>    <dd>Allows for ascending or descending sorting of the list.  The value maps properties to the sort direction (either <tt>asc</tt> for ascending or <tt>desc</tt> for descending).  Sortable properties are: <tt> dateCreated</tt>.</dd></dl>
+        *     <dt><tt>sorting</tt></dt>    <dd>Allows for ascending or descending sorting of the list.  The value maps properties to the sort direction (either <tt>asc</tt> for ascending or <tt>desc</tt> for descending).  Sortable properties are: <tt> id</tt><tt> label</tt>.</dd></dl>
         * @param     $authentication -  information used for the API call.  If no value is passed the global keys Simplify::public_key and Simplify::private_key are used.  <i>For backwards compatibility the public and private keys may be passed instead of the authentication object.</i>
-        * @return    ResourceList a ResourceList object that holds the list of Webhook objects and the total
-        *            number of Webhook objects available for the given criteria.
+        * @return    ResourceList a ResourceList object that holds the list of Tax objects and the total
+        *            number of Tax objects available for the given criteria.
         * @see       ResourceList
         */
-        static public function listWebhook($criteria = null, $authentication = null) {
+        static public function listTax($criteria = null, $authentication = null) {
 
             $args = func_get_args();
             $authentication = Simplify_PaymentsApi::buildAuthenticationObject($authentication, $args, 2);
 
-            $val = new Simplify_Webhook();
+            $val = new Simplify_Tax();
             $list = Simplify_PaymentsApi::listObject($val, $criteria, $authentication);
 
             return $list;
@@ -91,18 +98,18 @@ class Simplify_Webhook extends Simplify_Object {
 
 
         /**
-         * Retrieve a Simplify_Webhook object from the API
+         * Retrieve a Simplify_Tax object from the API
          *
-         * @param     string id  the id of the Webhook object to retrieve
+         * @param     string id  the id of the Tax object to retrieve
          * @param     $authentication -  information used for the API call.  If no value is passed the global keys Simplify::public_key and Simplify::private_key are used.  <i>For backwards compatibility the public and private keys may be passed instead of the authentication object.</i>
-         * @return    Webhook a Webhook object
+         * @return    Tax a Tax object
          */
-        static public function findWebhook($id, $authentication = null) {
+        static public function findTax($id, $authentication = null) {
 
             $args = func_get_args();
             $authentication = Simplify_PaymentsApi::buildAuthenticationObject($authentication, $args, 2);
 
-            $val = new Simplify_Webhook();
+            $val = new Simplify_Tax();
             $val->id = $id;
 
             $obj = Simplify_PaymentsApi::findObject($val, $authentication);
@@ -110,29 +117,10 @@ class Simplify_Webhook extends Simplify_Object {
             return $obj;
         }
 
-
-        /**
-         * Updates an Simplify_Webhook object.
-         *
-         * The properties that can be updated:
-         * <dl style="padding-left:10px;">
-         *     <dt><tt>url</tt></dt>    <dd>Endpoint URL <strong>required </strong></dd></dl>
-         * @param     $authentication -  information used for the API call.  If no value is passed the global keys Simplify::public_key and Simplify::private_key are used.  <i>For backwards compatibility the public and private keys may be passed instead of the authentication object.</i>
-         * @return    Webhook a Webhook object.
-         */
-        public function updateWebhook($authentication = null)  {
-
-            $args = func_get_args();
-            $authentication = Simplify_PaymentsApi::buildAuthenticationObject($authentication, $args, 1);
-
-            $object = Simplify_PaymentsApi::updateObject($this, $authentication);
-            return $object;
-        }
-
     /**
      * @ignore
      */
     public function getClazz() {
-        return "Webhook";
+        return "Tax";
     }
 }
