+++ src/Lib/Simplify/Simplify_TransactionReview.php
@@ -1,6 +1,12 @@
 <?php
+
+namespace Lib\Simplify;
+
+use Lib\Simplify\Simplify_Object;
+use Lib\Simplify\Simplify_PaymentsApi;
+use Lib\Simplify\Simplify_TransactionReview;
 /*
- * Copyright (c) 2013 - 2018 MasterCard International Incorporated
+ * Copyright (c) 2013 - 2019 MasterCard International Incorporated
  * All rights reserved.
  * 
  * Redistribution and use in source and binary forms, with or without modification, are 
@@ -27,28 +33,19 @@
  */
 
 
-class Simplify_Coupon extends Simplify_Object {
+class Simplify_TransactionReview extends Simplify_Object {
     /**
-     * Creates an Simplify_Coupon object
-     * @param     array $hash a map of parameters; valid keys are:<dl style="padding-left:10px;">
-     *     <dt><tt>amountOff</tt></dt>    <dd>Amount off of the price of the product in the smallest units of the currency of the merchant. While this field is optional, you must provide either amountOff or percentOff for a coupon. Example: 100 = $1.00 [min value: 1] </dd>
-     *     <dt><tt>couponCode</tt></dt>    <dd>Code that identifies the coupon to be used. [min length: 2] <strong>required </strong></dd>
-     *     <dt><tt>description</tt></dt>    <dd>A brief section that describes the coupon. </dd>
-     *     <dt><tt>durationInMonths</tt></dt>    <dd>DEPRECATED - Duration in months that the coupon will be applied after it has first been selected. [min value: 1, max value: 9999] </dd>
-     *     <dt><tt>endDate</tt></dt>    <dd>Last date of the coupon in UTC millis that the coupon can be applied to a subscription. This ends at 23:59:59 of the merchant timezone. </dd>
-     *     <dt><tt>maxRedemptions</tt></dt>    <dd>Maximum number of redemptions allowed for the coupon. A redemption is defined as when the coupon is applied to the subscription for the first time. [min value: 1] </dd>
-     *     <dt><tt>numTimesApplied</tt></dt>    <dd>The number of times a coupon will be applied on a customer's subscription. [min value: 1, max value: 9999] </dd>
-     *     <dt><tt>percentOff</tt></dt>    <dd>Percentage off of the price of the product. While this field is optional, you must provide either amountOff or percentOff for a coupon. The percent off is a whole number. [min value: 1, max value: 100] </dd>
-     *     <dt><tt>startDate</tt></dt>    <dd>First date of the coupon in UTC millis that the coupon can be applied to a subscription. This starts at midnight of the merchant timezone. <strong>required </strong></dd></dl>
+     * Creates an Simplify_TransactionReview object
+     * @param     array $hash a map of parameters; valid keys are:<dl style="padding-left:10px;"></dl>
      * @param     $authentication -  information used for the API call.  If no value is passed the global keys Simplify::public_key and Simplify::private_key are used.  <i>For backwards compatibility the public and private keys may be passed instead of the authentication object.<i/>
-     * @return    Coupon a Coupon object.
+     * @return    TransactionReview a TransactionReview object.
      */
-    static public function createCoupon($hash, $authentication = null) {
+    static public function createTransactionReview($hash, $authentication = null) {
 
         $args = func_get_args();
         $authentication = Simplify_PaymentsApi::buildAuthenticationObject($authentication, $args, 2);
 
-        $instance = new Simplify_Coupon();
+        $instance = new Simplify_TransactionReview();
         $instance->setAll($hash);
 
         $object = Simplify_PaymentsApi::createObject($instance, $authentication);
@@ -59,11 +56,11 @@ class Simplify_Coupon extends Simplify_Object {
 
 
        /**
-        * Deletes an Simplify_Coupon object.
+        * Deletes an Simplify_TransactionReview object.
         *
         * @param     $authentication -  information used for the API call.  If no value is passed the global keys Simplify::public_key and Simplify::private_key are used.  <i>For backwards compatibility the public and private keys may be passed instead of the authentication object.</i>
         */
-        public function deleteCoupon($authentication = null) {
+        public function deleteTransactionReview($authentication = null) {
 
             $args = func_get_args();
             $authentication = Simplify_PaymentsApi::buildAuthenticationObject($authentication, $args, 1);
@@ -75,23 +72,23 @@ class Simplify_Coupon extends Simplify_Object {
 
 
        /**
-        * Retrieve Simplify_Coupon objects.
+        * Retrieve Simplify_TransactionReview objects.
         * @param     array criteria a map of parameters; valid keys are:<dl style="padding-left:10px;">
-        *     <dt><tt>filter</tt></dt>    <dd><table class="filter_list"><tr><td>filter.id</td><td>Filter by the coupon Id</td></tr><tr><td>filter.text</td><td>Filter by the coupon code</td></tr><tr><td>filter.dateCreatedMin<sup>*</sup></td><td>Filter by the minimum created date you are searching for - Date in UTC millis</td></tr><tr><td>filter.dateCreatedMax<sup>*</sup></td><td>Filter by the maximum created date you are searching for - Date in UTC millis</td></tr><tr><td>filter.startDateMin</td><td>Filter by the minimum coupon start date you are searching for - Date in UTC millis</td></tr><tr><td>filter.startDateMax</td><td>Filter by the maximum coupon start date you are searching for - Date in UTC millis</td></tr><tr><td>filter.endDateMin</td><td>Filter by the minimum coupon end date you are searching for - Date in UTC millis</td></tr><tr><td>filter.endDateMax</td><td>Filter by the maximum coupon end date you are searching for - Date in UTC millis</td></tr></table><br><sup>*</sup>Use dateCreatedMin with dateCreatedMax in the same filter if you want to search between two created dates  </dd>
+        *     <dt><tt>filter</tt></dt>    <dd><table class="filter_list"><tr><td>filter.id</td><td>Filter by the Id of the transaction review</td></tr><tr><td>filter.label</td><td>Filter by the status of the transaction review</td></tr></table>  </dd>
         *     <dt><tt>max</tt></dt>    <dd>Allows up to a max of 50 list items to return. [min value: 0, max value: 50, default: 20]  </dd>
         *     <dt><tt>offset</tt></dt>    <dd>Used in paging of the list.  This is the start offset of the page. [min value: 0, default: 0]  </dd>
-        *     <dt><tt>sorting</tt></dt>    <dd>Allows for ascending or descending sorting of the list.  The value maps properties to the sort direction (either <tt>asc</tt> for ascending or <tt>desc</tt> for descending).  Sortable properties are: <tt> dateCreated</tt><tt> maxRedemptions</tt><tt> timesRedeemed</tt><tt> id</tt><tt> startDate</tt><tt> endDate</tt><tt> percentOff</tt><tt> couponCode</tt><tt> durationInMonths</tt><tt> numTimesApplied</tt><tt> amountOff</tt>.</dd></dl>
+        *     <dt><tt>sorting</tt></dt>    <dd>Allows for ascending or descending sorting of the list.  The value maps properties to the sort direction (either <tt>asc</tt> for ascending or <tt>desc</tt> for descending).  Sortable properties are: <tt> dateCreated</tt><tt> status</tt>.</dd></dl>
         * @param     $authentication -  information used for the API call.  If no value is passed the global keys Simplify::public_key and Simplify::private_key are used.  <i>For backwards compatibility the public and private keys may be passed instead of the authentication object.</i>
-        * @return    ResourceList a ResourceList object that holds the list of Coupon objects and the total
-        *            number of Coupon objects available for the given criteria.
+        * @return    ResourceList a ResourceList object that holds the list of TransactionReview objects and the total
+        *            number of TransactionReview objects available for the given criteria.
         * @see       ResourceList
         */
-        static public function listCoupon($criteria = null, $authentication = null) {
+        static public function listTransactionReview($criteria = null, $authentication = null) {
 
             $args = func_get_args();
             $authentication = Simplify_PaymentsApi::buildAuthenticationObject($authentication, $args, 2);
 
-            $val = new Simplify_Coupon();
+            $val = new Simplify_TransactionReview();
             $list = Simplify_PaymentsApi::listObject($val, $criteria, $authentication);
 
             return $list;
@@ -99,18 +96,18 @@ class Simplify_Coupon extends Simplify_Object {
 
 
         /**
-         * Retrieve a Simplify_Coupon object from the API
+         * Retrieve a Simplify_TransactionReview object from the API
          *
-         * @param     string id  the id of the Coupon object to retrieve
+         * @param     string id  the id of the TransactionReview object to retrieve
          * @param     $authentication -  information used for the API call.  If no value is passed the global keys Simplify::public_key and Simplify::private_key are used.  <i>For backwards compatibility the public and private keys may be passed instead of the authentication object.</i>
-         * @return    Coupon a Coupon object
+         * @return    TransactionReview a TransactionReview object
          */
-        static public function findCoupon($id, $authentication = null) {
+        static public function findTransactionReview($id, $authentication = null) {
 
             $args = func_get_args();
             $authentication = Simplify_PaymentsApi::buildAuthenticationObject($authentication, $args, 2);
 
-            $val = new Simplify_Coupon();
+            $val = new Simplify_TransactionReview();
             $val->id = $id;
 
             $obj = Simplify_PaymentsApi::findObject($val, $authentication);
@@ -120,16 +117,15 @@ class Simplify_Coupon extends Simplify_Object {
 
 
         /**
-         * Updates an Simplify_Coupon object.
+         * Updates an Simplify_TransactionReview object.
          *
          * The properties that can be updated:
          * <dl style="padding-left:10px;">
-         *     <dt><tt>endDate</tt></dt>    <dd>The ending date in UTC millis for the coupon. This must be after the starting date of the coupon. </dd>
-         *     <dt><tt>maxRedemptions</tt></dt>    <dd>Maximum number of redemptions allowed for the coupon. A redemption is defined as when the coupon is applied to the subscription for the first time. [min value: 1] </dd></dl>
+         *     <dt><tt>status</tt></dt>    <dd>Status of the transaction review. </dd></dl>
          * @param     $authentication -  information used for the API call.  If no value is passed the global keys Simplify::public_key and Simplify::private_key are used.  <i>For backwards compatibility the public and private keys may be passed instead of the authentication object.</i>
-         * @return    Coupon a Coupon object.
+         * @return    TransactionReview a TransactionReview object.
          */
-        public function updateCoupon($authentication = null)  {
+        public function updateTransactionReview($authentication = null)  {
 
             $args = func_get_args();
             $authentication = Simplify_PaymentsApi::buildAuthenticationObject($authentication, $args, 1);
@@ -142,6 +138,6 @@ class Simplify_Coupon extends Simplify_Object {
      * @ignore
      */
     public function getClazz() {
-        return "Coupon";
+        return "TransactionReview";
     }
 }
